+++ src/components/navigation/navbar/desktop/admin/Navbar.vue
@@ -17,7 +17,7 @@
 
 <script lang="ts">
     import BaseNavbar from "../common/BaseNavbar.vue";
-    import ManageAccountsMenu from "./administration/ManageAccountsMenu.vue";
+    import ManageAccountsMenu from "./administration/ManagementMenu.vue";
     import CourseMenu from "./courses/CourseMenu.vue";
     import ProfileMenu from "../common/profile/ProfileMenu.vue";
