+++ src/components/navigation/navbar/desktop/admin/administration/ManagementMenu.vue
@@ -15,7 +15,7 @@
 
 <script lang="ts">
     import BaseMenu from "../../common/BaseMenu.vue";
-    import MenuBody from "./ManageAccountsMenuBody.vue";
+    import MenuBody from "./ManagementMenuBody.vue";
 
     export default {
         components: {
