+++ src/components/navigation/navbar/mobile/admin/administration/ManagementMenuBody.vue
